refactor: 예외처리 구체화
