+++ jdbc/src/main/java/nextstep/jdbc/exception/DataTypeNotSupportedException.java
@@ -1,0 +1,10 @@
+package nextstep.jdbc.exception;
+
+public class DataTypeNotSupportedException extends RuntimeException{
+
+    private static final String MESSAGE = "지원하지 않는 데이터 타입입니다!";
+
+    public DataTypeNotSupportedException() {
+        super(MESSAGE);
+    }
+}
